pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "near-swarm"
version = "0.1.0"
description = "NEAR Protocol Swarm Intelligence Framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Jarrod Barnes", email = "jbarnes@near.org"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "aiohttp>=3.8.0",
    "python-dotenv>=0.19.0",
    "near-api-py>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=near_swarm"

[tool.pyright]
include = ["near_swarm/**"]
exclude = ["**/node_modules", "**/__pycache__"]
reportMissingImports = true
pythonVersion = "3.12"

[tool.pylance]
reportMissingImports = "warning" 